+++ substation-diagram/pom.xml
@@ -22,10 +22,11 @@
     <name>Substation diagram</name>
 
     <modules>
-        <module>substation-diagram-core</module>
-        <module>substation-diagram-view</module>
         <module>substation-diagram-cgmes</module>
+        <module>substation-diagram-core</module>
         <module>substation-diagram-util</module>
+        <module>substation-diagram-view</module>
+        <module>substation-diagram-view-app</module>
         <!--<module>substation-diagram-force-layout</module>-->
     </modules>
