+++ substation-diagram/substation-diagram-view-app/src/main/java/com/powsybl/substationdiagram/view/app/SubstationDiagramViewer.java
